physics update

+++ Octdoc/Code/graphics/application.cpp
@@ -6,7 +6,7 @@ namespace octdoc
 		void Application::Register(Graphics& graphics, Input& input)
 		{
 			graphics.setStartFunction([this](Graphics& g) { OnStart(g); });
-			graphics.setUpdateFunction([this](Graphics& g, float deltaTime) { OnUpdate(g, deltaTime); });
+			graphics.setUpdateFunction([this](Graphics& g, double deltaTime) { OnUpdate(g, deltaTime); });
 			input.setMouseMoveEvent([this](MouseMoveEvent& e) { OnMouseMove(e); });
 			input.setMouseLBtnDownEvent([this](MouseButtonEvent& e) { OnMouseLBtnDown(e); });
 			input.setMouseLBtnUpEvent([this](MouseButtonEvent& e) { OnMouseLBtnUp(e); });
